requirements.txt
Django~=2.2.4
django-model-utils==4.0.0
django-phonenumber-field
phonenumbers
django-crispy-forms==1.9.0
django-formtools==2.2